NEW: Данные разнесены по модулям, страницы сайта шаблонизированы
